% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPdist2}
\alias{fastPdist2}
\title{fastPdist2}
\usage{
fastPdist2(Ar, Br)
}
\arguments{
\item{Ar}{first matrix}

\item{Br}{second matrix}
}
\description{
Pairwise distance functions
}
\details{
\code{fastPdist} from http://blog.felixriedel.com/2013/05/pairwise-distances-in-r/
}

