% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{grid_design}
\alias{grid_design}
\title{Creates a grid design data frame}
\usage{
grid_design(data, grid_inds, reso = 20)
}
\arguments{
\item{data}{The original data frame}

\item{grid_inds}{The indices for the grid}

\item{reso}{The resolution of the grid}
}
\value{
A data frame with the same names, but variables in grid_inds replaced by a grid
}
\description{
Creates a grid design data frame
}

