% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/negatives.R
\name{logistic_truncation}
\alias{logistic_truncation}
\title{Logistic truncation}
\usage{
logistic_truncation(x, kmin, kmax)
}
\arguments{
\item{x}{Values to be truncated}

\item{k}{Truncation threshold}
}
\description{
Truncates values to [Kmin, Kmax] smoothly
}

