% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/negatives.R
\name{nw_surface}
\alias{nw_surface}
\title{Nonlinear surface}
\usage{
nw_surface(x, u, a, w, h = 1, handle = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{The data matrix}

\item{u}{Control points}

\item{a}{Value at control points}

\item{w}{Weights}

\item{h}{Bandwidth (default 1)}

\item{handle}{Return a function handle? If FALSE, returns the function values at x}

\item{normalize}{Adjust the function to have unit variance, zero mean, and zero regression coeffs}
}
\value{
A function handle if \code{handle = TRUE} and a vector if \code{handle = FALSE}
}
\description{
A synthetic function of the same class as fits produced by local regression
}

