% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{randn}
\alias{randn}
\title{MATLAB-style random normal matrix}
\usage{
randn(n, p = n, colnames = FALSE)
}
\arguments{
\item{n}{Dimension of the matrix}

\item{p}{Dimension of the matrix (default n)}

\item{colnames}{Whether to give colnames}
}
\description{
MATLAB-style random normal matrix
}

