% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{randu}
\alias{randu}
\title{Random uniform matrix}
\usage{
randu(n, p = n, center = FALSE, colnames = FALSE)
}
\arguments{
\item{n}{Dimension of the matrix}

\item{p}{Dimension of the matrix (default n)}

\item{center}{If true, generates [-1, 1] random variates}

\item{colnames}{Whether to give colnames}
}
\description{
Random uniform matrix
}

